reorder files, navigators and components
