+++ src/components/Button/Button.js
@@ -23,7 +23,6 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
     color: 'white'
   }
-
 })
 
 export default Button;
