+++ src/images/index.js
@@ -1,0 +1,4 @@
+export const ImageOne =  require('../../assets/images/item1.png'); 
+export const ImageTwo = require('../../assets/images/item2.png');
+export const ImageThree = require('../../assets/images/item3.png');
+export const ImageFour = require('../../assets/images/item4.png');
